Move ApiService and types to subdirectory
