+++ src/app/configurator/shared/options.type.ts
@@ -1,3 +1,3 @@
-import { OptionsApiData } from '../../shared/services/types/options-api-data.type';
+import { OptionsApiData } from '../../shared/services/api/types/options-api-data.type';
 
 export type Options = OptionsApiData;
