+++ src/app/configurator/summary-step/summary-step.component.ts
@@ -4,9 +4,9 @@ import { Observable, combineLatest, map } from 'rxjs';
 import { ErrorToastComponent } from '../../shared/components/error-toast/error-toast.component';
 import { ImageComponent } from '../../shared/components/image/image.component';
 import { TitleComponent } from '../../shared/components/title/title.component';
-import { Color } from '../../shared/services/types/color.type';
-import { Config } from '../../shared/services/types/config.type';
-import { Model } from '../../shared/services/types/model.type';
+import { Color } from '../../shared/services/api/types/color.type';
+import { Config } from '../../shared/services/api/types/config.type';
+import { Model } from '../../shared/services/api/types/model.type';
 import { ConfiguratorService } from '../configurator.service';
 import { Error } from '../shared/error.type';
 import { Image } from '../shared/image.type';
