+++ src/app/shared/services/api/types/api-endpoints.type.ts
