+++ src/app/shared/services/api/types/config.type.ts
