+++ src/app/shared/services/api/types/model.type.ts
