+++ src/app/shared/services/api/types/options-api-data.type.ts
