+++ src/app/shared/services/api/types/options-api-response.type.ts
